# back_end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Recherche d'emploi",
        "description": "Documentation de l'API Laravel pour la plateforme de recherche d'emploi",
        "contact": {
            "email": "hariznabil663@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur principal"
        }
    ],
    "paths": {
...
        "/jobs": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Get all jobs",
                "operationId": "61095dd3fdb7b25e7daa0084b396ca6e",
                "responses": {
                    "200": {
                        "description": "List of jobs"
                    }
                }
            },
            "post": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Create a new job",
                "operationId": "724b21b194916208376d71437be61a8b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "company_name",
                                    "location",
                                    "employment_type"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "company_name": {
                                        "type": "string"
                                    },
                                    "location": {
                                        "type": "string"
                                    },
                                    "employment_type": {
                                        "type": "string",
                                        "enum": [
                                            "full-time",
                                            "part-time",
                                            "contract",
                                            "internship"
                                        ]
                                    },
                                    "salary": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Job posted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/jobs/{id}": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Get a single job",
                "operationId": "40e7eaa66d57253caa249e6f6fc06763",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job details"
                    }
                }
            },
            "put": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Update a job",
                "operationId": "73ddd3048c72f5dbb787a4733069f5d1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "company_name": {
                                        "type": "string"
                                    },
                                    "location": {
                                        "type": "string"
                                    },
                                    "employment_type": {
                                        "type": "string",
                                        "enum": [
                                            "full-time",
                                            "part-time",
                                            "contract",
                                            "internship"
                                        ]
                                    },
                                    "salary": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Job updatedd successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Delete a job",
                "operationId": "7d6ee44631cf0dc1175069d51532c9b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the job to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Job not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/jobs/search": {
            "post": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Search jobs",
                "operationId": "64479b9803ef5aed151ffad67a4b9459",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "location": {
                                        "type": "string"
                                    },
                                    "company_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Filtered jobs"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}